+++ src/components/Projects/ProjectsPreviewSection.tsx
@@ -1,13 +1,27 @@
 import Image from 'next/image';
 import { useState } from 'react';
 
+import { useProjectPageContext } from '@/context/projectPage.context';
+
 import { ProjectCard } from './ProjectCard/ProjectCard';
-import { getProjectsOnSite } from './Projects.utils';
+import { getProjectsOnSite, PROJECTS_ARRAY } from './Projects.utils';
 
 export const ProjectPreviewSection = () => {
+  const { filterHashtags, filterStatuses } = useProjectPageContext();
   const [siteNumber, setSiteNumber] = useState<number>(0);
   const maxProjectsOnThePage = 3;
-  const projectsOnForSites = getProjectsOnSite(maxProjectsOnThePage);
+  const filteredProjectsArray = PROJECTS_ARRAY.filter(
+    (project) =>
+      !!project.hashtags.find((projectHashtag) =>
+        filterHashtags.find((filterHashtag) => filterHashtag === projectHashtag),
+      ),
+  );
+  const projectsOnForSites = getProjectsOnSite({
+    array: filteredProjectsArray.length > 0 ? filteredProjectsArray : PROJECTS_ARRAY,
+    arrayDivider: maxProjectsOnThePage,
+  });
+
+  console.log(filteredProjectsArray);
 
   return (
     <section className="z-10">
